third_party/WebKit/LayoutTests/inspector/editor/text-editor-smart-braces-expected.txt

This test checks text editor smart braces functionality.


Running: testTypeBraceSequence
({[|]})

Running: testBraceOverride
({[]})|

Running: testQuotesToCloseStringLiterals
'Hello"'|
